# Apply JEC/JER smearing via correctionlib