This PR implements new calibrators for applying the jet energy corrections/jet energy resolution smearing and the associated uncertainties via correctionlib.

The implementation closely follows the previous one using `coffea.jetmet_tools`, but reads in the corrections and applies them to the jets using the fast vectorized evaluators provided by `correctionlib`. The new implementation is now the default, but the old one is kept in `calibration/jets_coffea.py` for reference and validation purposes.

To use the new implementation, an external file containing the correction tables needs to be added to the config. All JEC levels and uncertainties, as well as the JER values, scale factors, and scale factor variations in correctionlib format, are now provided centrally via a single gzipped JSON file, which must be supplied via a new `jet_jerc` entry in `cfg.x.external_files`. The `jec`, `junc`, `jer` and `jersf` external file dicts are not used by the new implementation and can be removed.

The `jec` and `jer` auxiliary config entries can be kept mostly as they are, but the `levels` key can now be a list with a compound correction level (typically `L1L2L3Res`) as its only element instead of the full list of correction levels. This compound correction applies all the required correction levels in the right order, taking care to propagate changes in the jet p<sub>T</sub> at each stage. The individual correction levels used as the starting point for the Type-I MET propagation must be configured via the `cfg.x.jec.levels_for_type1_met` key. The `source` key is no longer needed and can be removed.

To validate the new implementation, I compared some of the corrected jet and MET quantities to the old implementation and found only negligible differences:
```python
events_old = self[jets_coffea](events, **kwargs)
events_new = self[jets](events, **kwargs)

def compare(route, comp_func=lambda a, b: abs(a - b)):
    from columnflow.columnar_util import Route
    route = Route(route)
    old = route.apply(events_old)
    new = route.apply(events_new)
    comp = comp_func(old, new)
    return {
        "min":  round(ak.min(comp),  5),
        "mean": round(ak.mean(comp), 5),
        "max":  round(ak.max(comp),  5),
    }

>>> compare("Jet.pt")
... {'min': 0.0, 'mean': 0.0, 'max': 5e-05}
>>> compare("MET.pt")
... {'min': 0.0, 'mean': 2e-05, 'max': 0.00023}
>>> compare("MET.phi")
... {'min': 0.0, 'mean': 0.0, 'max': 0.0001}

>>> compare("Jet.pt_jec_Total_up")
... {'min': 0.0, 'mean': 1e-05, 'max': 0.00017}
>>> compare("MET.pt_jec_Total_up")
... {'min': 0.0, 'mean': 2e-05, 'max': 0.00032}

>>> compare("Jet.pt_jer_up")
... {'min': 0.0, 'mean': 0.0, 'max': 5e-05}
>>> compare("MET.pt_jer_up")
... {'min': 0.0, 'mean': 1e-05, 'max': 7e-05}
```


The correctionlib implementation appears to be about 2.7 (!) times faster than the one using `coffea.jetmet_tools`:

```bash
# coffea.jetmet_tools
iterate through 197113 events in 4 chunks ...
handling chunk 3
calibrator 'jets_coffea' took 14.107 seconds  # first chunk also loads the correction files
handling chunk 0
calibrator 'jets_coffea' took 7.374 seconds
handling chunk 1
calibrator 'jets_coffea' took 7.303 seconds
handling chunk 2
calibrator 'jets_coffea' took 7.183 seconds
done (took 38 seconds)
event processing in loop body took 36.04 seconds

# correctionlib
iterate through 197113 events in 4 chunks ...
handling chunk 3
calibrator 'jets_correctionlib' took 2.680 seconds
handling chunk 0
calibrator 'jets_correctionlib' took 2.732 seconds
handling chunk 1
calibrator 'jets_correctionlib' took 2.671 seconds
handling chunk 2
calibrator 'jets_correctionlib' took 2.662 seconds
done (took 12.69 seconds)
event processing in loop body took 10.82 seconds
```